frontier-audio-jarvis/ai-service/requirements.txt
fastapi
uvicorn
openai
python-dotenv
websockets
pydub
requests
